fix: align rewards API with review feedback.

+++ src/main/java/com/visa/rewards/dto/MonthlyReward.java
@@ -1,9 +1,10 @@
 package com.visa.rewards.dto;
 
+import java.math.BigDecimal;
 import java.time.YearMonth;
 
 public record MonthlyReward(
         YearMonth month,
-        long points
+        BigDecimal points
 ) {
 }
