+++ src/main/java/com/visa/rewards/dto/RewardResponse.java
@@ -1,10 +1,11 @@
 package com.visa.rewards.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record RewardResponse(
         String customerId,
         List<MonthlyReward> monthlyRewards,
-        long totalPoints
+        BigDecimal totalPoints
 ) {
 }
